pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fleetdb-mcp-server"
version = "0.1.0"
description = "MCP server exposing a PostgreSQL fleet database with a two-phase confirmation gate for writes"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Yashwin Vasanth Srinivasan" }
]
keywords = ["mcp", "model-context-protocol", "postgresql", "agentic-ai", "fleet", "mobility"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp[cli]>=1.27.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "sqlparse>=0.5.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
fleetdb-mcp = "fleetdb_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fleetdb_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
